* add live inference mode rgb and flow

+++ live_inference.py
@@ -25,12 +25,12 @@ _MIX_WEIGHT_OF_FLOW = 0.5
 _LOG_ROOT = 'output'
 
 #Run Video Examples
-#video_path = './data/example/_Archery_g01_c01.avi'
+video_path = './data/example/_Archery_g01_c01.avi'
 #video_path = './data/example/_BoxingPunchingBag_g01_c01.avi' 
 #video_path = './data/example/_JugglingBalls_g01_c02.avi' 
 
 #Run camera 
-video_path = 0
+#video_path = 0
 
 DATA_DIR = 'data/live'
 prediction_print=['none',0,0]
@@ -80,12 +80,18 @@ def main(dataset, mode, variant, workers, inference):
     
     # start a new session and restore the fine-tuned model
     sess.run(tf.global_variables_initializer())
-    
-    rgb_saver.restore(sess, _CHECKPOINT_PATHS['rgb'])
-    flow_saver.restore(sess, _CHECKPOINT_PATHS['flow'])
+    if mode in ['rgb', 'mixed']:
+        rgb_saver.restore(sess, _CHECKPOINT_PATHS['rgb'])
+    if mode in ['flow', 'mixed']:
+        flow_saver.restore(sess, _CHECKPOINT_PATHS['flow'])
 
     # Here we start the inference
     print('----Here we start!----', end="\r\n")
+    print('Mode is set to:',mode, end="\r\n") 
+    print('Dataset is set to:',dataset, end="\r\n") 
+    print('FlowNet2 variant is set to:',variant, end="\r\n") 
+    print('Inference frame is set to:',inference, end="\r\n") 
+    print('Workers is set to:',workers, end="\r\n") 
     print('Output writes to '+ log_dir, end="\r\n")
   
     start_time = time.time()
@@ -95,12 +101,12 @@ def main(dataset, mode, variant, workers, inference):
     if not os.path.exists(DATA_DIR+'/flow'):
         os.system("mkdir "+DATA_DIR+'/flow')
 
-    read_stream(variant,prob_predictions,fc_out, rgb_fc_out, flow_fc_out, rgb_saver, flow_saver, rgb_holder, flow_holder, label_map, sess, inference, workers)
+    read_stream(variant,mode,prob_predictions,fc_out, rgb_fc_out, flow_fc_out, rgb_saver, flow_saver, rgb_holder, flow_holder, label_map, sess, inference, workers)
 
     print('Infer Input in: '+str(time.time() - start_time), end="\r\n")
     sess.close()
 
-def read_stream(variant, prob_predictions, fc_out, rgb_fc_out, flow_fc_out, rgb_saver, flow_saver, rgb_holder, flow_holder, label_map, sess, inference_frames, workers):
+def read_stream(variant,mode, prob_predictions, fc_out, rgb_fc_out, flow_fc_out, rgb_saver, flow_saver, rgb_holder, flow_holder, label_map, sess, inference_frames, workers):
     pooling_executor = ThreadPoolExecutor(max_workers=workers)
     future_threads = []
     cv2.namedWindow("Preview")
@@ -127,8 +133,9 @@ def read_stream(variant, prob_predictions, fc_out, rgb_fc_out, flow_fc_out, rgb_
         flow_path = DATA_DIR+"/flow/"+str(imagelist_number+1)+"/frame{:06d}.npy".format(count)
         flow_paths.append(flow_path)
         if count != 0 and count%inference_frames == 0:
-            print('Calculate Flows for Part '+str(int(imagelist_number))+'...', end="\r\n")
-            future_threads.append(pooling_executor.submit(generate_flow_and_predict,variant,DATA_DIR,flow_paths,img_paths,img_frames,str(imagelist_number),prob_predictions, fc_out, rgb_fc_out, flow_fc_out, rgb_saver, flow_saver, rgb_holder, flow_holder, label_map, sess))
+            if mode in ['flow', 'mixed']:
+                print('Calculate Flows for Part '+str(int(imagelist_number))+'...', end="\r\n")
+            future_threads.append(pooling_executor.submit(generate_flow_and_predict,variant,mode,DATA_DIR,flow_paths,img_paths,img_frames,str(imagelist_number),prob_predictions, fc_out, rgb_fc_out, flow_fc_out, rgb_saver, flow_saver, rgb_holder, flow_holder, label_map, sess))
             print('Added data of Part '+str(int(imagelist_number))+' in pipeline...', end="\r\n")
             img_frames = []
             img_paths = []
@@ -144,26 +151,38 @@ def read_stream(variant, prob_predictions, fc_out, rgb_fc_out, flow_fc_out, rgb_
     wait(future_threads)
     print('All threads finished!', end="\r\n")
 
-def generate_flow_and_predict(variant,DATA_DIR,flow_paths,img_paths,img_frames,imagelist_number,prob_predictions,fc_out, rgb_fc_out, flow_fc_out, rgb_saver, flow_saver, rgb_holder, flow_holder, label_map, sess):
-    write_flownet_io_files(DATA_DIR,flow_paths,img_paths,imagelist_number)
-    flow_frames = generate_flows(DATA_DIR,variant,imagelist_number)
-    predict_action(img_frames, flow_frames,imagelist_number,prob_predictions, fc_out, rgb_fc_out, flow_fc_out, rgb_saver, flow_saver, rgb_holder, flow_holder, label_map, sess)
-    clean_directory(DATA_DIR,imagelist_number,img_paths)
+def generate_flow_and_predict(variant,mode,DATA_DIR,flow_paths,img_paths,img_frames,imagelist_number,prob_predictions,fc_out, rgb_fc_out, flow_fc_out, rgb_saver, flow_saver, rgb_holder, flow_holder, label_map, sess):
+    flow_frames = []
+    if mode in ['flow', 'mixed']:
+        write_flownet_io_files(DATA_DIR,flow_paths,img_paths,imagelist_number)
+        flow_frames = generate_flows(DATA_DIR,variant,imagelist_number)
+    predict_action(mode,img_frames, flow_frames,imagelist_number,prob_predictions, fc_out, rgb_fc_out, flow_fc_out, rgb_saver, flow_saver, rgb_holder, flow_holder, label_map, sess)
+    clean_directory(DATA_DIR,mode,imagelist_number,img_paths)
     
-def predict_action(img_frames, flow_frames, imagelist_number,prob_predictions, fc_out, rgb_fc_out, flow_fc_out, rgb_saver, flow_saver, rgb_holder, flow_holder, label_map, sess):
-    rgb_clip = np.array(img_frames)  
-    rgb_clip = rgb_clip/255 
-    input_rgb = rgb_clip[np.newaxis, :, :, :, :]
+def predict_action(mode,img_frames, flow_frames, imagelist_number,prob_predictions, fc_out, rgb_fc_out, flow_fc_out, rgb_saver, flow_saver, rgb_holder, flow_holder, label_map, sess):
+    if mode in ['rgb', 'mixed']:
+        rgb_clip = np.array(img_frames)  
+        rgb_clip = rgb_clip/255 
+        input_rgb = rgb_clip[np.newaxis, :, :, :, :]
      
-    flow_clip = np.array(flow_frames)
-    flow_clip = 2*(flow_clip/255)-1
-    input_flow = flow_clip[np.newaxis, :, :, :, :]
-      
-    out_prob_predictions ,logit_predictions, curr_rgb_fc_data, curr_flow_fc_data = sess.run(
-        [prob_predictions,fc_out, rgb_fc_out, flow_fc_out],
-        feed_dict={rgb_holder: input_rgb, flow_holder: input_flow,})
+    if mode in ['flow', 'mixed']:
+        flow_clip = np.array(flow_frames)
+        flow_clip = 2*(flow_clip/255)-1
+        input_flow = flow_clip[np.newaxis, :, :, :, :]
+
+    if mode in ['rgb']:  
+        out_prob_predictions ,logit_predictions, curr_rgb_fc_data,  = sess.run(
+            [prob_predictions,fc_out, rgb_fc_out],
+            feed_dict={rgb_holder: input_rgb,})
+    elif mode in ['flow']:  
+        out_prob_predictions ,logit_predictions, curr_flow_fc_data,  = sess.run(
+            [prob_predictions,fc_out, flow_fc_out],
+            feed_dict={flow_holder: input_flow,})
+    elif mode in ['mixed']:  
+        out_prob_predictions ,logit_predictions, curr_rgb_fc_data, curr_flow_fc_data = sess.run(
+            [prob_predictions,fc_out, rgb_fc_out, flow_fc_out],
+            feed_dict={rgb_holder: input_rgb, flow_holder: input_flow,})
         
-    # np.argmax(predictions[0]),predictions[0, np.argmax(predictions, axis=1)[0]]
     print('Predict Part '+imagelist_number+ ' with Class ' + trans_label(np.argmax(out_prob_predictions[0]), label_map) + ' with probability ' + str(max(out_prob_predictions[0])), end="\r\n")
     update_predictions_inframe(imagelist_number,trans_label(np.argmax(out_prob_predictions[0]), label_map),max(out_prob_predictions[0]))
 
@@ -175,14 +194,17 @@ def update_predictions_inframe(imagelist_number,prediction,confidence):
         prediction_print[2] = confidence
 
 def build_model(mode,dataset):
+    rgb_holder = None
     if mode in ['rgb', 'mixed']: 
         rgb_holder = tf.placeholder(
             tf.float32, [None, None, _FRAME_SIZE, _FRAME_SIZE, _CHANNEL['rgb']])
+    flow_holder = None
     if mode in ['flow', 'mixed']:
         flow_holder = tf.placeholder(
             tf.float32, [None, None, _FRAME_SIZE, _FRAME_SIZE, _CHANNEL['flow']])
 
     # insert the model
+    rgb_fc_out = None
     if mode in ['rgb', 'mixed']:
         with tf.variable_scope(_SCOPE['rgb']):
             rgb_model = i3d.InceptionI3d(
@@ -192,7 +214,7 @@ def build_model(mode,dataset):
             rgb_logits_dropout = tf.nn.dropout(rgb_logits, 1)
             rgb_fc_out = tf.layers.dense(
                 rgb_logits_dropout, _CLASS_NUM[dataset], use_bias=True)
-                
+    flow_fc_out = None            
     if mode in ['flow', 'mixed']:
         with tf.variable_scope(_SCOPE['flow']):
             flow_model = i3d.InceptionI3d(
@@ -205,6 +227,7 @@ def build_model(mode,dataset):
 
     # construct two separate feature map and saver(rgb_saver,flow_saver)
     variable_map = {}
+    rgb_saver = None
     if mode in ['rgb', 'mixed']:
         for variable in tf.global_variables():
             tmp = variable.name.split('/')
@@ -212,6 +235,7 @@ def build_model(mode,dataset):
                 variable_map[variable.name.replace(':0', '')] = variable
         rgb_saver = tf.train.Saver(var_list=variable_map)
     variable_map = {}
+    flow_saver = None
     if mode in ['flow', 'mixed']:
         for variable in tf.global_variables():
             tmp = variable.name.split('/')
@@ -220,13 +244,13 @@ def build_model(mode,dataset):
         flow_saver = tf.train.Saver(var_list=variable_map, reshape=True)
 
     # Edited Version by AlexHu
-    if mode == 'rgb':
+    if mode in ['rgb']:
         fc_out = rgb_fc_out
         prob_predictions = tf.nn.softmax(fc_out)
-    if mode == 'flow':
+    elif mode in ['flow']:
         fc_out = flow_fc_out
         prob_predictions = tf.nn.softmax(fc_out)
-    if mode == 'mixed':
+    elif mode in ['mixed']:
         fc_out = _MIX_WEIGHT_OF_RGB * rgb_fc_out + _MIX_WEIGHT_OF_FLOW * flow_fc_out
         prob_predictions = tf.nn.softmax(fc_out)
         
@@ -278,15 +302,16 @@ def generate_flows(DATA_DIR,variant,imagelist_number):
         flow_frames.append(transform_data(uv,mode='flow'))
     return flow_frames
     
-def clean_directory(DATA_DIR,imagelist_number,img_path):
+def clean_directory(DATA_DIR,mode,imagelist_number,img_path):
     img_str_delete = ""
     for img in img_path[:-1]:
         img_str_delete = img_str_delete + ' '+img
     os.system("rm "+img_str_delete)
-    os.system("rm -r "+DATA_DIR+'/flow/'+imagelist_number)  
-    os.system("rm "+DATA_DIR+'/images'+imagelist_number+'_1.txt')
-    os.system("rm "+DATA_DIR+'/images'+imagelist_number+'_2.txt')
-    os.system("rm "+DATA_DIR+'/output'+imagelist_number+'.txt')
+    if mode in ['flow', 'mixed']:
+        os.system("rm -r "+DATA_DIR+'/flow/'+imagelist_number)  
+        os.system("rm "+DATA_DIR+'/images'+imagelist_number+'_1.txt')
+        os.system("rm "+DATA_DIR+'/images'+imagelist_number+'_2.txt')
+        os.system("rm "+DATA_DIR+'/output'+imagelist_number+'.txt')
 
 if __name__ == '__main__':
     description = 'Test Finetuned I3D Model'
